Remove circular Dependency warning with SimpleList sub-components - rollup does static analysis on modules and reports things like circular   dependencies. As this is a general bad coding practise, it is better   to refactor to a Parent->Child dependency chain. - moved the SimpleList Context to an External Dependency to enable the   SimpleList and it's child components to compose Context without   having a circular dependency.
